Updated UI and added fucntionality to 'Next' and 'Prev' buttons
updated search function to allow for pagination.
Added a listbox to search results to let users to select specific results to get reviews for that result.
Removed the option to search the different files since listboxes made more since.
Changed the search button to reflect this.
Added a function to update the titles of the reviews field when selecting a specific company to display it's name.
Added functionality to the 'next' and 'prev' buttons.

Next will be to add the 'summarize' functionality