Some minor fixes
Moved Clear button code to the top.
Fixed a bug with restricting how many characters you could right into the token_limit field.
Changed the variable name for what that char limit is for more clairty
Added a bit to titleText to include which LLM is being used when the results get saved to file.